% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demoTimesTwoRcpp}
\alias{demoTimesTwoRcpp}
\title{Demo of basic timesTwo function}
\usage{
demoTimesTwoRcpp()
}
\description{
A simple driver for the timesTwoRcpp function. No parameters.
}
\examples{
demoTimesTwoRcpp()

}
